refactored tests to remove code duplications when running the same test on different architectures
